Change PNG for SVG and refactor some redundant lines
